Big rework regarding threading.
The main work is now done on a separate thread, and the main thread
handles rendering / input. This solves some issues with the VRAM window
and any further windows being created, and may allow gbcc to build on
OSX (though this is untested). Also, specifying a max turbo speed with
`-t` will now also enable sound in turbo mode.