Add support for Vsync and interlacing.
This allows various minor effects from the Game Boy to be reproduced,
such as the "wobbling" of shapes when the screen scrolls.

Vsync is still considered experimental, as it hasn't seen much testing,
and there is no re-shifting of the pitch so audio is pitched up very
slightly (by about ~0.5%). It was necessary to double the length of the
audio buffer to 2048 samples, or about 21ms @ 96kHz.

Some other minor changes are also added, mostly out of laziness :)

  - Slight rewording of the man page
  - Shader toggle key moved to 'o', to allow Vsync to have 'v'
  - Nanosecond constants changed to be of type unsigned long rather than
    just unsigned
  - Minor cleanup of code in window.c