Rework interlacing and frameblending shader code.
Previously, interlacing (and part of frameblending) was being done on
the CPU, before any colour corrections from the various other shaders.
This led to incorrect colours, as the operations of darkening / mixing a
pixel and colour-correcting it are not commutative. It also meant that
interlacing couldn't be enabled if Vsync wasn't also on, as there would
be horrible flickering of the screen.

These are now handled in shaders, and performed after any
colour-correction. This fixes the colours, simplifies the code a small
amount, and allows interlacing to be enabled without Vsync.

N.B. There is an implicit assumption that the maximum texture size
supported by the system is at least as big as the scaled GBC screen, as
the largest required by OpenGL is only 2048x2048. I think this is pretty
much guaranteed, and will only be a problem if the screen is more than
144/160 * 2048 ~= 1840 pixels in height (and the graphics card doesn't
support bigger than 2048x2048).