Fix some APU and audio behaviour.
- Change audio data format to signed 16 bit.
- Don't clock channel 1 & 2 duty timers after powering on until they've
  been triggered once.

This fixes the audio in Perfect Dark, which would otherwise be silent
most of the time. This is because the game produces audio by altering
the master output level of the Game Boy, to produce arbitrary waveform
audio.

Previously, the DAC was treated as having an output range of [0, 1] (in
arbitrary units), whereas it really has a range of [-1, 1].
A digital channel output of 0 therefore produces an analogue output of
-1; changing the master volume will then change the signal being output,
even if the *digital* signal is constant at 0.

This change also somewhat reproduces the audio pops heard whenever a
channel DAC is enabled or disabled.