Fix potential crash on Android.
The menu now prints the current palette name, which is part of the core
struct. On Android, there is no guarantee that the core has been
initialised by the time we first call `gbcc_menu_update()`, so this
palette name could be a null (or otherwise invalid) pointer. We
therefore need to check if the core is initialised when we update the
menu.