Potential fix for random hang.
Some games, such as Pokemon Crystal, would randomly hang after a few
minutes (or even seconds) of gameplay on Android. This seemed to be
happen only when pressing buttons, and occured much more frequently on
Android than desktop.

It *seems* that this was caused by directly setting the interrupt flag
bit from gbcc_input_process_key, which happens outside the main thread,
and could therefore lead to impossible behaviour. This is now changed to
set a boolean flag, which is used to set the appropriate bit of the IF
register on the main emulation thread.