Stop exiting on invalid opcode.
Previously, encountering an invalid opcode would cause gbcc to
immediately exit. This is fine on desktop, but on android gbcc is used
as a shared library, and this lead to crashing. Now, the error flag is
set, and the emulation thread quits immediately on error - it's up to
the front end to detect this.