Update savestate loading.
Previously, tmp_core would be allocated on the stack, but it's a bit big
for that (~60kB) so now it's heap allocated. Additionally, the logic
used to go like:

  1. Load the savestate into a temporary core struct.
  2. Copy the values across to the current core.
  3. Fix any pointers.

This meant that another thread (e.g. the rendering thread) could try to
read an invalid pointer after step 2. Steps 2 & 3 are now swapped, so
that shouldn't be an issue anymore.